BREAKING: changed base image to Ubuntu
The aws cli v2 does not support Alpine linux (yet). To solve the problem,
I changed the base image to Ubuntu 18.04 (the latest LTS release
at the time of the commit). We were already getting a tagged version
of Terraform via curl, and the difference in size for the resulting
image is overshadowed by the size of the aws cli install itself.

We could optimize for image size at a later point, for example by
removing unnecessary files created by the `apt update` op.