changed the name of the repository
Updated the documentation files and CI configuration to point to the new name of
the repository: "cloud-ci".

The context has changed dramatically around the use of this image, going from
simply supporting the execution of the `terraform-aws-eks` module and its
dependencies on local binaries to being a more global image to run CI/CD
pipelines. As such it didn't make much sense to keep the name "terraform-eks"
for this image. "cloud-ci" might be too generic but at least it allows for a
wider array of use-cases.