updated various dependencies
Updated base version of Ubuntu to the latest LTS release, and updated the
versions for the apps installed via apt.

Updated the versions of various applications installed on the image. Terraform
is left untouched because it can cause issues where some people would be forced
to update their version of Terraform locally. Terraform will be changed in a
separate release after coordinating with the different development teams
internally.